pc - update for CS48 S20
