+++ .gitignore
@@ -1,3 +1,11 @@
+# secret properties files
+
+secrets-*.properties
+
+# Credentials downloaded from Google
+
+credential*.json
+
 # Emacs/vim
 
 *~
@@ -15,10 +23,6 @@
 
 .DS_Store
 
-# CS56 specific: Secrets files
-
-localhost.json
-heroku.json
 
 # Maven (from: https://github.com/github/gitignore/blob/master/Maven.gitignore)
