+++ convertGoogleJson.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 
 # This file takes the values in heroku.json, and sets those values
 # on a remote app.
@@ -14,17 +14,15 @@ from pprint import pprint
 import sys
 
 
-if len(sys.argv)!=3:
+if len(sys.argv)!=2:
   print("Converts credentials.json from Google OAuth to format needed")
-  print("  for Spring Boot SPRING_APPLICATION_JSON format")
-  print("Usage: "+sys.argv[0]+" input output")
+  print("  for Spring Boot properties files")
+  print("Usage: "+sys.argv[0]+" input ")
   print("   where ")
   print("      input is filename of a credentials.json file downloaded from Google OAuth configuration")
-  print("      output is either localhost.json or heroku.json")
   sys.exit(1)
 
 input_filename = sys.argv[1]
-output_filename = sys.argv[2]
 
 
 try:
@@ -39,21 +37,11 @@ except Exception as e:
   print(type(e),e.args)
   sys.exit(3)
 
-output_dict = {}
-
 client_id_key = "spring.security.oauth2.client.registration.google.client-id"
 client_secret_key = "spring.security.oauth2.client.registration.google.client-secret"
 
+client_id = input_dict["web"]["client_id"]
+client_secret = input_dict["web"]["client_secret"]
 
-output_dict[client_id_key] = input_dict["web"]["client_id"]
-output_dict[client_secret_key] = input_dict["web"]["client_secret"]
-
-
-try:
-  with open(output_filename,'w') as output_file:
-    json.dump(output_dict, output_file)
-except Exception as e:
-  print("Error writing to ", output_filename)
-  print(type(e),e.args)
-  sys.exit(4)
-
+print(f"{client_id_key}: {client_id}")
+print(f"{client_secret_key}: {client_secret}")
