+++ secrets-localhost.properties.SAMPLE
@@ -1,2 +1,2 @@
-spring.security.oauth2.client.registration.github.client-id=REPLACE-ME
-spring.security.oauth2.client.registration.github.client-secret=REPLACE-ME
+spring.security.oauth2.client.registration.google.client-id=REPLACE-ME
+spring.security.oauth2.client.registration.google.client-secret=REPLACE-ME
