+++ setHerokuEnv.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+heroku config:set HEROKU_PROPERTIES="$(cat secrets-heroku.properties)" "$@"
