+++ src/main/java/edu/ucsb/cs48/s20/demo/Config.java
@@ -1,4 +1,4 @@
-package edu.ucsb.cs56.w20.lab07;
+package edu.ucsb.cs48.s20.demo;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
