+++ src/main/java/edu/ucsb/cs48/s20/demo/advice/AuthControllerAdvice.java
@@ -1,11 +1,11 @@
-package edu.ucsb.cs56.w20.lab07.advice;
+package edu.ucsb.cs48.s20.demo.advice;
 
 import org.springframework.web.bind.annotation.ModelAttribute;
 
-import edu.ucsb.cs56.w20.lab07.controllers.AppUsersController;
-import edu.ucsb.cs56.w20.lab07.entities.AppUser;
-import edu.ucsb.cs56.w20.lab07.repositories.AppUserRepository;
-import edu.ucsb.cs56.w20.lab07.services.MembershipService;
+import edu.ucsb.cs48.s20.demo.controllers.AppUsersController;
+import edu.ucsb.cs48.s20.demo.entities.AppUser;
+import edu.ucsb.cs48.s20.demo.repositories.AppUserRepository;
+import edu.ucsb.cs48.s20.demo.services.MembershipService;
 
 import java.util.List;
