+++ src/main/java/edu/ucsb/cs48/s20/demo/repositories/AdminRepository.java
@@ -1,11 +1,11 @@
-package edu.ucsb.cs56.w20.lab07.repositories;
+package edu.ucsb.cs48.s20.demo.repositories;
 
 import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import edu.ucsb.cs56.w20.lab07.entities.Admin;
+import edu.ucsb.cs48.s20.demo.entities.Admin;
 
 @Repository
 public interface AdminRepository extends CrudRepository<Admin, Long> {
