+++ src/main/java/edu/ucsb/cs48/s20/demo/services/MembershipService.java
@@ -1,4 +1,4 @@
-package edu.ucsb.cs56.w20.lab07.services;
+package edu.ucsb.cs48.s20.demo.services;
 
 import java.util.List;
