+++ src/main/resources/application.properties
@@ -1,4 +1,6 @@
 spring.jpa.show-sql=true
 spring.jpa.hibernate.ddl-auto=update
-app.admin.emails=phtcon@ucsb.edu,scottpchow@ucsb.edu,zsisco@ucsb.edu,pingyuan@ucsb.edu
+app.admin.emails=phtcon@ucsb.edu,scottpchow@ucsb.edu
 app.member.hosted-domain=ucsb.edu
+
+@environment-properties@
