+++ src/main/resources/templates/bootstrap/bootstrap_footer.html
@@ -1,7 +1,7 @@
 <!-- Footer -->
 <footer class="page-footer font-small secondary-color pt-4">
     <p>
-        This page is a class project of <a href="https://ucsb-cs56.github.io">UCSB's CMPSC 56</a>.
+        This page is a class project of <a href="https://ucsb-cs48.github.io">UCSB's CMPSC 48</a>.
 
     </p>
