+++ src/main/resources/templates/bootstrap/bootstrap_nav_header.html
@@ -3,7 +3,7 @@
 
 <nav class="navbar navbar-expand-lg navbar-light bg-light">
 
-    <a class="navbar-brand" href="/">lab07</a>
+    <a class="navbar-brand" href="/">demo</a>
 
     <div class="collapse navbar-collapse" id="navbarTogglerDemo03">
         <ul class="navbar-nav mr-auto mt-2 mt-lg-0">
