+++ src/main/resources/templates/bootstrap/title.html
@@ -1,1 +1,1 @@
-<title>lab07-CS56</title>
+<title>CS48 demo</title>
