+++ src/test/java/edu/ucsb/cs48/s20/demo/controllers/HomePageTest.java
@@ -1,4 +1,4 @@
-package edu.ucsb.cs56.w20.lab07.controllers;
+package edu.ucsb.cs48.s20.demo.controllers;
 
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
@@ -13,7 +13,7 @@ import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 
-import edu.ucsb.cs56.w20.lab07.advice.AuthControllerAdvice;
+import edu.ucsb.cs48.s20.demo.advice.AuthControllerAdvice;
 
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.xpath;
 
@@ -44,6 +44,6 @@ public class HomePageTest {
     public void getHomePage_hasCorrectTitle() throws Exception {
         mvc.perform(MockMvcRequestBuilders.get("/").accept(MediaType.TEXT_HTML)).andExpect(status().isOk())
                 .andExpect(xpath("//title").exists())
-                .andExpect(xpath("//title").string("lab07-CS56"));
+                .andExpect(xpath("//title").string("CS48 demo"));
     }
 }
